Rename 'types' modules to 'responses'
