+++ bee-inx/src/node/mod.rs
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod requests;
-pub mod types;
+pub mod responses;
 
 use futures::stream::{Stream, StreamExt};
 
-pub use self::{requests::*, types::*};
+pub use self::{requests::*, responses::*};
 use crate::{
     client::{try_from_inx_type, Inx},
     error::Error,
